+++ DLVN-SDK/src/main/java/com/edoctor/dlvn_sdk/helper/NotificationHelper.kt
@@ -49,6 +49,11 @@ object NotificationHelper {
         val intent = Intent(context, mainActivityClass).apply {
             flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
         }
+        val accept = Intent(context, mainActivityClass).apply {
+            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
+            action = "ACCEPT_CALL_FROM_QUIT_STATE"
+        }
+        val acceptIntent: PendingIntent = PendingIntent.getActivity(context, 0, accept, PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT)
         val fullScreenIntent: PendingIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT)
 
         val decline = Intent(context, CallActionReceiver::class.java)
@@ -63,15 +68,16 @@ object NotificationHelper {
             .setCategory(NotificationCompat.CATEGORY_CALL)
             .setPriority(NotificationCompat.PRIORITY_HIGH)
             .setContentIntent(fullScreenIntent)
+            .setSilent(true)
             .setFullScreenIntent(fullScreenIntent, true)
             .addAction(R.drawable.end_call_24, context.getString(R.string.incoming_decline_label), declineIntent)
-            .addAction(R.drawable.accept_call_24, context.getString(R.string.incoming_accept_label), fullScreenIntent)
+            .addAction(R.drawable.accept_call_24, context.getString(R.string.incoming_accept_label), acceptIntent)
             .setAutoCancel(true)
             .setOngoing(false)
 
-//        val mediaPlayer = MediaPlayer.create(context, R.raw.ringing)
-//        mediaPlayer.isLooping = true
-//        mediaPlayer.start()
+        val mediaPlayer = MediaPlayer.create(context, R.raw.ringing)
+        mediaPlayer.isLooping = true
+        mediaPlayer.start()
 
         notificationManager!!.notify( /*notification ID*/1, builder.build())
     }