+++ DLVN-SDK/src/main/java/com/edoctor/dlvn_sdk/sendbirdCall/CallManager.kt
@@ -36,6 +36,7 @@ class CallManager {
     var mContext: Context? = null
     var totalTime: Int = 0
     private var apiService: ApiService? = null
+    var acceptCallFromQuitState: Boolean = false
     var acceptCallSetting: AcceptSetting? = AcceptSetting()
     var appointmentDetail: AppointmentDetailInfo? = AppointmentDetailInfo()