+++ DLVN-SDK/src/main/java/com/edoctor/dlvn_sdk/sendbirdCall/WebViewCallActivity.kt
@@ -2,8 +2,8 @@ package com.edoctor.dlvn_sdk.sendbirdCall
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.util.Log
 import android.view.View
+import android.view.WindowManager
 import android.widget.ImageButton
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
@@ -35,6 +35,8 @@ class WebViewCallActivity: AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_webview)
+        window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+
         webView.domain = getChatRoomUrl()
         if (savedInstanceState == null) {
             webView.hideLoading = true